diff, Tagger.pm, Files.pm, Plain.pm, CVS.pm, BK.pm: fix for bug #2714744
"Tagger & Plain inefficient"

When files need to be handed over to an external tool, a temporary copy was made and the tool operated in this copy. In the case of plain files, this is inefficient since the file is already available in the tree.
Method tmpfile changed to realfilename where the name of the plain file is returned or the name of the copy. New method releaserealfilename is called when the file is no longer needed. This does nothing for plain files and deletes temp copy otherwise.
Notes: 1- Git.pm not updated because support is broken in CPAN.
       2- BK.pm blindly updated by guess (current maintainer has no BK copy nor experience).

(cvs update 4 Apr 2012, part 1)